exchanges, encoding/json, currency: answer review feedback and fix ticker last prices
exchanges/deribit: incremental_ticker decodes each change onto the instrument's stored state,
so a field a change leaves out keeps its value and one it nulls, as it does the day's range
once the 24 hour window empties, is cleared. The websocket ticker maps every instrument kind as
REST does, and all three ticker paths report open interest.

exchanges/huobi, kraken, poloniex: tickers survive a null or one-sided book, kraken's websocket
reads the 24 hour index as REST does, and poloniex records the requested pair.

exchanges/huobi, bithumb: every ticker path sets Last, which the engine reads for price events
and unrealised PNL, and bithumb's websocket ticker takes it from closePrice rather than the
previous day's close.

exchanges/binanceus: WsOrderUpdateData gains WorkingTime for "W", which matched the bool "w"
case-insensitively and failed the update of every order resting on the book.
WebsocketAggregateTradeStream gains BestMatchPrice for "M" for the same reason.

exchanges, currency: integers move off int, to uint64 where the provider never sends a negative
and to int64, commented, where it can, and field names spell acronyms in one case. kraken's
futures order responses get named types, and forex timestamps decode as types.Time. Tests use
httptest.NewTestServer and set the exchange name per test where they touch a global store.

encoding/json: document the sonic decoder at a read boundary, and give the real reason
AllowDuplicateNames is needed.

+++ currency/forexprovider/currencyconverterapi/currencyconverterapi_types.go
@@ -32,7 +32,7 @@ type CurrencyConverter struct {
 
 // Error stores the error message
 type Error struct {
-	Status int    `json:"status"`
+	Status int64  `json:"status"` // Signed because the error format is undocumented
 	Error  string `json:"error"`
 }
