+++ exchanges/coinut/coinut_websocket.go
@@ -760,7 +760,7 @@ func (e *Exchange) wsSubmitOrders(ctx context.Context, orders []WsSubmitOrderPar
 				Price:         orders[i].Price,
 				Side:          orders[i].Side.String(),
 				InstrumentID:  e.instrumentMap.LookupID(curr.String()),
-				ClientOrderID: i + 1,
+				ClientOrderID: uint64(i) + 1,
 			})
 	}
