+++ exchanges/gateio/gateio_websocket_request_types.go
@@ -83,7 +83,7 @@ type WebsocketOrderResponse struct {
 	GTDiscount                bool          `json:"gt_discount"`
 	RebatedFee                types.Number  `json:"rebated_fee"`
 	RebatedFeeCurrency        currency.Code `json:"rebated_fee_currency"`
-	SelfTradePreventionID     int           `json:"stp_id"`
+	SelfTradePreventionID     uint64        `json:"stp_id"`
 	SelfTradePreventionAction string        `json:"stp_act"`
 	AverageDealPrice          types.Number  `json:"avg_deal_price"`
 	Label                     string        `json:"label"`
